requirements.txt
flask
flask-sqlalchemy
flask-login
werkzeug
python-dotenv
gunicorn
pywebpush
flask-apscheduler
flask-wtf
flask-talisman
flask-limiter
flask-migrate
tzdata
webuntis
cryptography

# Google Drive Integration
google-api-python-client
google-auth-httplib2
google-auth-oauthlib

# PDF Processing & OCR
pdfplumber
PyPDF2
pillow

